Conflicts:
	client/components/sublist/sublist.css
	client/components/sublist/sublist.html

+++ client/app/statuseditor/statuseditor.controller.js
@@ -39,15 +39,17 @@ angular.module('umm3601ursamajorApp')
         };
 
         $scope.deleteStatus = function(item){
-            console.log("Deleting status: " + item.strict);
-            $http.delete('/api/statuss/' + item._id).success(function(){
-                $scope.statusArray.splice($scope.statusArray.indexOf(item), 1);
-            });
-            var threshold = item.priority;
-            for(var j = 0; j < status.length; j++) {
-                if($scope.statusArray[j].priority != 1 || $scope.statusArray[j].priority != 15) {
-                    if ($scope.statusArray[j].priority > threshold) {
-                        $scope.statusArray[j].priority--;
+            var r = confirm("Are you sure you want to delete this status? All statuses will with this status will need to be changed.")
+            if(r == true) {
+                $http.delete('/api/statuss/' + item._id).success(function () {
+                    $scope.statusArray.splice($scope.statusArray.indexOf(item), 1);
+                });
+                var threshold = item.priority;
+                for (var j = 0; j < status.length; j++) {
+                    if ($scope.statusArray[j].priority != 1 || $scope.statusArray[j].priority != 15) {
+                        if ($scope.statusArray[j].priority > threshold) {
+                            $scope.statusArray[j].priority--;
+                        }
                     }
                 }
             }
@@ -66,73 +68,55 @@ angular.module('umm3601ursamajorApp')
         };
 
         $scope.addStatus = function() {
-            $http.post('/api/statuss/',
-            {   strict: "Default Status",
-                color: {red: 0, green: 0, blue: 0, alpha: 1},
-                emailSubject: "",
-                emailBody: "",
-                priority: $scope.findEmptyPriority($scope.statusArray)
-            }).success(function (){
-                    console.log("Succesfully added new status")
-                    $scope.getStatuses();
-                });
+            var r = confirm("Are you sure you want to add a status?")
+            if(r == true) {
+                $http.post('/api/statuss/',
+                    {   strict: "Default Status",
+                        color: {red: 0, green: 0, blue: 0, alpha: 1},
+                        emailSubject: "",
+                        emailBody: "",
+                        priority: $scope.findEmptyPriority($scope.statusArray)
+                    }).success(function () {
+                        console.log("Succesfully added new status")
+                        $scope.getStatuses();
+                    });
+            }
         };
 
 
         $scope.requiredStatus = function(status){
-            return(status.priority == 1 || status.priority == 15);
+            return(status.priority == -15 || status.priority == 15);
         };
-
 
         $scope.submitChanges = function(status) {
-            var x = $scope.statusArray.indexOf(status);
-            var strict = "";
             var r = confirm("Are you sure you want to edit this status?");
-            $http.get('/api/statuss/' + $scope.statusArray[x]._id).success(function(oldStatus) {
-                console.log("this should come first");
-                strict  = oldStatus.strict;
-                if (r){
-                    $http.patch('/api/statuss/' + $scope.statusArray[x]._id,
-                        {
-                            strict: $scope.statusArray[x].strict,
-                            color: $scope.statusArray[x].color,
-                            emailSubject: $scope.statusArray[x].emailSubject,
-                            emailBody: $scope.statusArray[x].emailBody
-                        }
-                    ).success(function () {
-                            console.log("this should come second");
-                            for(var j = 0; j < $scope.submissions.length; j++){
-                                if($scope.submissions[j].status.strict == strict){
-                                    console.log("things were detected to be different");
-                                    $scope.submissions[j].status.strict = $scope.statusArray[x].strict;
-                                    $http.patch('/api/submissions/' + $scope.submissions[j]._id, {
-                                        status: {strict: $scope.statusArray[x].strict, text: $scope.submissions[j].status.text}
-                                    })
-                                }
-                            }
-                            $location.path('/admin');
-                        })
+            var conflict = false;
+            var priorityOne = false
+            var x = $scope.statusArray.indexOf(status);
+            if (r){
+                for(var i=0; i < $scope.statusArray.length; i++) {
+                    if($scope.statusArray[i].priority == status.priority){
+                        conflict = true;
+                    }
+                    if(status.priority == 1){
+                        priorityOne = true;
+                    }
                 }
-            });
+                    if(!conflict && !priorityOne) {
+                        $http.patch('/api/statuss/' + $scope.statusArray[x]._id,
+                            {
+                                strict: $scope.statusArray[x].strict,
+                                color: $scope.statusArray[x].color,
+                                emailSubject: $scope.statusArray[x].emailSubject,
+                                emailBody: $scope.statusArray[x].emailBody,
+                                priority: $scope.statusArray[x].priority
+                            }
+                        ).success(function () {
+                                $location.path('/admin');
+                            })
+                    } else {
+                        alert("There already exists a status with this priority.")
+                    }
+                }
         };
-
-
-        //OLD
-//        $scope.submitChanges = function() {
-//            var r = confirm("Are you sure you want to edit this status?")
-//            if (r){
-//                for(var x = 0; x < $scope.statusArray.length; x++){
-//                    $http.put('/api/statuss/' + $scope.statusArray[x]._id,
-//                        {
-//                            strict: $scope.statusArray[x].strict,
-//                            color: $scope.statusArray[x].color,
-//                            emailSubject: $scope.statusArray[x].emailSubject,
-//                            emailBody: $scope.statusArray[x].emailBody
-//                        }
-//                    ).success(function () {
-//                            $location.path('/admin');
-//                        })
-//                }
-//            }
-//        };
     });
